+++ crates/veryl/src/cmd_publish.rs
@@ -1,5 +1,5 @@
 use crate::OptPublish;
-use log::debug;
+use log::{debug, warn};
 use miette::{IntoDiagnostic, Result};
 use std::time::Instant;
 use veryl_metadata::Metadata;
@@ -16,14 +16,17 @@ impl CmdPublish {
     pub fn exec(&self, metadata: &Metadata) -> Result<bool> {
         let now = Instant::now();
 
-        if let Some(kind) = self.opt.bump {
-            metadata.bump_version(kind.into()).into_diagnostic()?;
-        }
-
         let mut metadata = metadata.clone();
 
+        if let Some(kind) = self.opt.bump {
+            metadata.bump_version(kind.into()).into_diagnostic()?;
+            if !metadata.publish.bump_commit {
+                warn!("Please git add and commit: Veryl.toml");
+                return Ok(true);
+            }
+        }
+
         metadata.publish()?;
-        metadata.save_pubdata()?;
 
         let elapsed_time = now.elapsed();
         debug!("Elapsed time ({} milliseconds)", elapsed_time.as_millis());