+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Added] lint for naming convention [#116](https://github.com/dalance/veryl/issues/116)
 * [Added] `veryl publish` command
+* [Changed] version of dependency becomes mandatory
 
 ## [v0.3.4](https://github.com/dalance/veryl/compare/v0.3.3...v0.3.4) - 2023-02-08